requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
black==23.3.0
google==3.0.0
isort==5.12.0
python-dotenv==1.0.0
protobuf==4.22.3
requests==2.28.2